Get the global package variable sbc_config.

The configuration is a list of items.

cbc_quantiles

a named list of site-specific quantile functions for each CBC of interest

cbc_abnormals

a named list of site-specific functions that flag values as abnormal or not

cbc_vars

a list of names of CBC variables of interest, i.e. specific values of BASE_NAME

census_locations

a character vector locations that need to be used for modeling

surgery_services

a character vector of OR services that need to be used for modeling

c0

a value to use in training for c0

history_window

how many days of history to use in training, default 200

penalty_factor

the penalty factor in training, default 15

start

the day when the model evaluation begins, default 10

initial_collection_data

the number of units to collect on the prediction day, a day after, and another day after, i.e. a 3-vector

initial_expiry_data

the number of units expiring a day after prediction begins and one day after that, i.e. a 2-vector

data_folder

full path of location of raw data files

log_folder

full path of where logs should go, must exist

model_update_frequency

how often to update the model, default 7 days

lag_window

number of previous days to average in smoothing, default 7 days

cbc_filename_prefix

a character expression describing the prefix of daily CBC file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

census_filename_prefix

a character expression describing the prefix of daily census file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

surgery_filename_prefix

a character expression describing the prefix of daily surgery file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

transfusion_filename_prefix

a character expression describing the prefix of daily transfusion file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

output_filename_prefix

a character expression describing the prefix of daily output file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

log_filename_prefix

a character expression describing the prefix of log file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

inventory_filename_prefix

a character expression describing the prefix of the inventory file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

get_SBC_config()

Value

the value of package global sbc_config.

Examples

get_SBC_config()

#> $cbc_quantiles
#> $cbc_quantiles$HGB
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <bytecode: 0x7fa0f29f4908>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$LYMAB
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <bytecode: 0x7fa0f29f7ee8>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$MCH
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <bytecode: 0x7fa0f29f7820>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$MCHC
#> function (x) 
#> 32
#> <bytecode: 0x7fa0f29f6ba8>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$MCV
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <bytecode: 0x7fa0f29f65f8>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$PLT
#> function (x) 
#> 150
#> <bytecode: 0x7fa0f29f8630>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$RBC
#> function (x) 
#> 4.4
#> <bytecode: 0x7fa0f29fbf90>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$RDW
#> function (x) 
#> 12
#> <bytecode: 0x7fa0f29fb9a8>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$WBC
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <bytecode: 0x7fa0f29fb0b0>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_quantiles$HCT
#> function (x) 
#> 40
#> <bytecode: 0x7fa0f29fa780>
#> <environment: 0x7fa0f29ea070>
#> 
#> 
#> $cbc_abnormals
#> $cbc_abnormals$HGB
#> function (x) 
#> x < 9
#> <bytecode: 0x7fa0f29fa208>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$LYMAB
#> function (x) 
#> x < 0.69
#> <bytecode: 0x7fa0f29fdac0>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$MCH
#> function (x) 
#> x < 29
#> <bytecode: 0x7fa0f29fd120>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$MCHC
#> function (x) 
#> x < 32
#> <bytecode: 0x7fa0f29fc898>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$MCV
#> function (x) 
#> x < 87
#> <bytecode: 0x7fa0f29fc278>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$PLT
#> function (x) 
#> x < 150
#> <bytecode: 0x7fa0f29ff548>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$RBC
#> function (x) 
#> x < 4.4
#> <bytecode: 0x7fa0f29fef60>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$RDW
#> function (x) 
#> x < 11.5
#> <bytecode: 0x7fa0f2a03fc8>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$WBC
#> function (x) 
#> x > 10.3
#> <bytecode: 0x7fa0f2a02c50>
#> <environment: 0x7fa0f29ea070>
#> 
#> $cbc_abnormals$HCT
#> function (x) 
#> x < 34
#> <bytecode: 0x7fa0f2a025f8>
#> <environment: 0x7fa0f29ea070>
#> 
#> 
#> $census_locations
#>  [1] "B1"          "B2"          "B3"          "C1"          "C2"         
#>  [6] "C3"          "E2-ICU"      "E1"          "J2"          "J4"         
#> [11] "J5"          "J6"          "J7"          "K4"          "K5"         
#> [16] "K6"          "K7"          "L4"          "L5"          "L6"         
#> [21] "L7"          "M4"          "M5"          "M6"          "M7"         
#> [26] "VCP 1 WEST"  "VCP 2 WEST"  "VCP 2 NORTH" "VCP 3 WEST"  "VCP CCU 1"  
#> [31] "VCP CCU 2"   "VCP NICU"    "VCP NURSERY"
#> 
#> $surgery_services
#> [1] "Transplant"               "Cardiac"                 
#> [3] "Bone Marrow Transplant"   "Neurosurgery"            
#> [5] "Thoracic"                 "Vascular"                
#> [7] "Hepatology"               "General"                 
#> [9] "Interventional Radiology"
#> 
#> $c0
#> [1] 10
#> 
#> $history_window
#> [1] 100
#> 
#> $penalty_factor
#> [1] 15
#> 
#> $prediction_bias
#> [1] 10
#> 
#> $start
#> [1] 10
#> 
#> $initial_collection_data
#> [1] 60 60 60
#> 
#> $initial_expiry_data
#> [1] 0 0
#> 
#> $min_inventory
#> [1] 0
#> 
#> $data_folder
#> [1] "home/app/platelet_data_sample"
#> 
#> $database_path
#> [1] "home/app/database.duckdb"
#> 
#> $cbc_filename_prefix
#> [1] "Hospital_Daily_CBC"
#> 
#> $census_filename_prefix
#> [1] "Hospital_Daily_Census"
#> 
#> $transfusion_filename_prefix
#> [1] "Hospital_Daily_Transfusion"
#> 
#> $inventory_filename_prefix
#> [1] "Daily_Morning_Inventory"
#> 
#> $surgery_filename_prefix
#> [1] "Hospital_Daily_Surgery"
#> 
#> $log_filename_prefix
#> [1] "SBCpip_%s.json"
#> 
#> $model_update_frequency
#> [1] 7
#> 
#> $lag_window
#> [1] 7
#> 
#> $l1_bounds
#>  [1] 60 58 56 54 52 50 48 46 44 42 40 38 36 34 32 30 28 26 24 22 20 18 16 14 12
#> [26] 10  8  6  4  2  0
#> 
#> $lag_bounds
#> [1] -1 10
#> 
#> $org_cbc_cols
#> [1] "ORDER_PROC_ID" "BASE_NAME"     "RESULT_TIME"   "ORD_VALUE"    
#> 
#> $org_census_cols
#> [1] "PAT_ID"        "LOCATION_NAME" "LOCATION_DT"  
#> 
#> $org_surgery_cols
#> [1] "LOG_ID"           "OR_SERVICE"       "SURGERY_DATE"     "FIRST_SCHED_DATE"
#> [5] "CASE_CLASS"      
#> 
#> $org_transfusion_cols
#> [1] "DIN"             "Product Code"    "Type"            "Issue Date/Time"
#> 
#> $org_inventory_cols
#> [1] "Inv. ID"        "Type"           "Days to Expire" "Exp. Date"     
#> [5] "Exp. Time"     
#> 
#> $cbc_vars
#> [1] "HGB"   "LYMAB" "MCH"   "MCHC"  "MCV"   "PLT"   "RBC"   "RDW"   "WBC"  
#> 
#> $log_folder
#> [1] "home/app/platelet_logs"
#>