Read a single cbc file data and return it, as is

Read a single cbc file data and return it, as is

read_one_cbc_file(
  filename,
  cbc_abnormals,
  cbc_vars,
  org_cols = c("ORDER_PROC_ID", "BASE_NAME", "RESULT_TIME", "ORD_VALUE")
)

Arguments

filename
the fully qualified path of the file

cbc_abnormals,
a named list of functions of a single vector returning TRUE for abnormal values

cbc_vars,
the names of values to include; others are excluded

org_cols,
the names of columns at the target organization (specified in data mapping)

Value

a list of four items, filename, raw_data (tibble), report a list of missing values tibble and a summary tibble, cbc_data (tibble)

Examples

config <- SBCpip::get_SBC_config()
filename <- system.file("extdata", "platelet_data_sample", 
                        "Hospital_Daily_CBC2020-05-02-fake.csv",
                        package = "SBCpip")
cbc_data <- read_one_cbc_file(filename, config$cbc_abnormals, config$cbc_vars)

#> {"timestamp": "2021-11-18T17:15:02-0800", "log_lvl": "INFO", "log_msg": "Processing "}

#> Error: '' does not exist in current working directory ('/Users/naras/GitHub/SBCpip/docs/reference').