Get the global package variable sbc_config.

The configuration is a list of items.

cbc_quantiles

a named list of site-specific quantile functions for each CBC of interest

cbc_abnormals

a named list of site-specific functions that flag values as abnormal or not

cbc_vars

a list of names of CBC variables of interest, i.e. specific values of BASE_NAME

census_locations

a character vector locations that need to be used for modeling

c0

a value to use in training for c0

min_inventory

a value to use as offset to ensure a minimum threshold

history_window

how many days of history to use in training, default 200

penalty_factor

the penalty factor in training, default 15

start

the day when the model evaluation begins, default 10

initial_expiry_data

the number of units expiring a day after prediction begins and one day after that, i.e. a 2-vector

initial_collection_data

the number of units to collect on the prediction day, a day after, and another day after, i.e. a 3-vector

data_folder

full path of location of raw data files

report_folder

full path of where reports should go, must exist

output_folder

full path of where processed output data should go, must exist

log_folder

full path of where logs should go, must exist

model_update_frequency

how often to update the model, default 7 days

cbc_filename_prefix

a character expression describing the prefix of daily CBC file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

census_filename_prefix

a character expression describing the prefix of daily census file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

transfusion_filename_prefix

a character expression describing the prefix of daily transfusion file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

output_filename_prefix

a character expression describing the prefix of daily output file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

log_filename_prefix

a character expression describing the prefix of log file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

inventory_filename_prefix

a character expression describing the prefix of the inventory file name, this is combined with the date is substituted with the date in YYYY-dd-mm format

get_SBC_config()

Value

the value of package global sbc_config.

Examples

get_SBC_config()

#> $cbc_quantiles
#> $cbc_quantiles$HGB
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$LYMAB
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$MCH
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$MCHC
#> function (x) 
#> 32
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$MCV
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$PLT
#> function (x) 
#> 150
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$RBC
#> function (x) 
#> 4.4
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$RDW
#> function (x) 
#> 12
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$WBC
#> function (x) 
#> quantile(x, probs = 0.25, na.rm = TRUE)
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_quantiles$HCT
#> function (x) 
#> 40
#> <environment: 0x7fd46936a4b8>
#> 
#> 
#> $cbc_abnormals
#> $cbc_abnormals$HGB
#> function (x) 
#> x < 9
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$LYMAB
#> function (x) 
#> x < 0.69
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$MCH
#> function (x) 
#> x < 29
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$MCHC
#> function (x) 
#> x < 32
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$MCV
#> function (x) 
#> x < 87
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$PLT
#> function (x) 
#> x < 150
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$RBC
#> function (x) 
#> x < 4.4
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$RDW
#> function (x) 
#> x < 11.5
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$WBC
#> function (x) 
#> x > 10.3
#> <environment: 0x7fd46936a4b8>
#> 
#> $cbc_abnormals$HCT
#> function (x) 
#> x < 34
#> <environment: 0x7fd46936a4b8>
#> 
#> 
#> $census_locations
#>  [1] "B1"                   "B2"                   "B3"                  
#>  [4] "C1"                   "C2"                   "C3"                  
#>  [7] "CAPR XFER OVERFL"     "CATH PACU"            "CDU-CLIN DEC UNIT"   
#> [10] "D1CC"                 "D1CS"                 "D2"                  
#> [13] "D3"                   "DGR"                  "E1"                  
#> [16] "E2-ICU"               "E29-ICU"              "E3"                  
#> [19] "EMERGENCY DEPARTMENT" "F3"                   "FGR"                 
#> [22] "G1"                   "G2P"                  "G2S"                 
#> [25] "H1"                   "H2"                  
#> 
#> $c0
#> [1] 30
#> 
#> $min_inventory
#> [1] 30
#> 
#> $history_window
#> [1] 200
#> 
#> $penalty_factor
#> [1] 15
#> 
#> $start
#> [1] 10
#> 
#> $initial_collection_data
#> [1] 60 60 60
#> 
#> $initial_expiry_data
#> [1] 0 0
#> 
#> $data_folder
#> [1] "E:/platelet_predict_daily_data"
#> 
#> $cbc_filename_prefix
#> [1] "LAB-BB-CSRP-CBC_Daily%s-"
#> 
#> $census_filename_prefix
#> [1] "LAB-BB-CSRP-Census_Daily%s-"
#> 
#> $transfusion_filename_prefix
#> [1] "LAB-BB-CSRP-Transfused Product Report_Daily%s-"
#> 
#> $inventory_filename_prefix
#> [1] "Daily_Product_Inventory_Report_Morning_To_Folder%s-"
#> 
#> $output_filename_prefix
#> [1] "pip-output-%s.RDS"
#> 
#> $log_filename_prefix
#> [1] "SBCpip_%s.json"
#> 
#> $model_update_frequency
#> [1] 7
#> 
#> $cbc_vars
#> [1] "HGB"   "LYMAB" "MCH"   "MCHC"  "MCV"   "PLT"   "RBC"   "RDW"   "WBC"  
#> 
#> $report_folder
#> [1] "E:/Blood_Center_Reports"
#> 
#> $output_folder
#> [1] "E:/Blood_Center_Output"
#> 
#> $log_folder
#> [1] "E:/Blood_Center_Logs"
#>